♻️ : 채팅 관련 상태 ChatArea로 이동
